coworking_project/requirements/base.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
django-filter==24.2
drf-spectacular==0.27.2
celery==5.4.0
redis==5.0.6
stripe==10.5.0
pillow==10.3.0
django-ratelimit==4.1.0
whitenoise==6.7.0